# docs/swagger.yaml
basePath: /content
definitions:
...
  RewardClaim:
    properties:
      date_created:
        type: string
      date_updated:
        type: string
      description:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/RewardClaimItem'
        type: array
      org_id:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  RewardClaimItem:
    properties:
      amount:
        type: integer
      inventory_id:
        type: string
      reward_type:
        type: string
    type: object
...
  model.RewardQuantityState:
    properties:
      claimable_quantity:
        type: integer
      grantable_quantity:
        type: integer
      reward_type:
        type: string
    type: object
host: localhost
info:
  contact: {}
  description: RoRewards Building Block API Documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Rewards Building Block API
  version: 1.0.6
paths:
...
  /int/reward_history:
    post:
      consumes:
      - application/json
      description: Gets reward quantity stats for the current moment
      operationId: InternalGetRewardStats
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.RewardQuantityState'
            type: array
      security:
      - InternalApiAuth: []
      tags:
      - Internal
  /user/balance:
    get:
      description: Retrieves balance for each user's wallet
      operationId: GetUserBalance
      responses:
        "200":
          description: ""
      security:
      - UserAuth: []
      tags:
      - Client
  /user/claim:
    post:
      consumes:
      - application/json
      description: Create a new claim user claim
      operationId: CreateUserRewardClaim
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RewardClaim'
      security:
      - AdminUserAuth: []
      tags:
      - Client
  /user/claims:
    get:
      consumes:
      - application/json
      description: Gets user claims
      operationId: GetUserRewardClaim
      parameters:
      - description: status
        in: query
        name: status
        type: string
      - description: limit - limit the result
        in: query
        name: limit
        type: string
      - description: offset
        in: query
        name: offset
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/RewardClaim'
            type: array
      security:
      - AdminUserAuth: []
      tags:
      - Client
  /user/history:
    get:
      description: Retrieves the wallet history
      operationId: GetUserRewardsHistory
      parameters:
      - description: reward_type - filter by reward_type
        in: query
        name: reward_type
        type: string
      - description: code - filter by code
        in: query
        name: code
        type: string
      - description: filter by building_block
        in: query
        name: building_block
        type: string
      - description: limit - limit the result
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: ""
      security:
      - UserAuth: []
      tags:
      - Client
...
schemes:
- https
securityDefinitions:
  AdminGroupAuth:
    in: header
    name: GROUP
    type: apiKey
  AdminUserAuth:
    in: header (add Bearer prefix to the Authorization value)
    name: Authorization
    type: apiKey
  InternalApiAuth:
    in: header (add INTERNAL-API-KEY with correct value as a header)
    name: Authorization
    type: apiKey
swagger: "2.0"